refactor amp/bento-lightbox-gallery
fix lightbox gallery deps
